+++ simple-props/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 4.2 - TBD
+## 4.2 - 2017-10-01
 ### Changes
 * Added loading from `InputStream`
 * Added saving to `OutputStream`